Feature: Add Atomic Design Pattern with Absolute imports
