+++ App.tsx
@@ -3,7 +3,7 @@ import { withExpoSnack } from 'nativewind'
 import { NavigationContainer } from '@react-navigation/native'
 import { createNativeStackNavigator } from '@react-navigation/native-stack'
 
-import HomeScreen from './screens/HomeScreen'
+import HomeScreen from '@src/screens/HomeScreen'
 
 const Stack = createNativeStackNavigator()
