+++ babel.config.js
@@ -2,6 +2,17 @@ module.exports = function (api) {
   api.cache(true)
   return {
     presets: ['babel-preset-expo'],
-    plugins: ['nativewind/babel']
+    plugins: [
+      'nativewind/babel',
+      [
+        'module-resolver',
+        {
+          root: ['.'],
+          alias: {
+            '@src': './src'
+          }
+        }
+      ]
+    ]
   }
 }
