+++ package.json
@@ -19,6 +19,7 @@
     "@react-navigation/native-stack": "^6.9.8",
     "expo": "~47.0.12",
     "expo-status-bar": "~1.4.2",
+    "expo-updates": "~0.15.6",
     "nativewind": "^2.0.11",
     "react": "18.1.0",
     "react-dom": "18.1.0",
@@ -27,14 +28,14 @@
     "react-native-safe-area-context": "4.4.1",
     "react-native-screens": "~3.18.0",
     "react-native-svg": "13.4.0",
-    "react-native-web": "^0.18.10",
-    "expo-updates": "~0.15.6"
+    "react-native-web": "^0.18.10"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
     "@types/react": "~18.0.24",
     "@types/react-native": "~0.70.6",
     "@typescript-eslint/eslint-plugin": "^5.48.1",
+    "babel-plugin-module-resolver": "^5.0.0",
     "eslint": "^8.32.0",
     "eslint-config-prettier": "^8.6.0",
     "eslint-config-standard-with-typescript": "^27.0.1",
@@ -42,12 +43,12 @@
     "eslint-plugin-n": "^15.6.1",
     "eslint-plugin-promise": "^6.1.1",
     "eslint-plugin-react": "^7.32.0",
+    "husky": "^8.0.0",
     "prettier": "^2.8.2",
     "prettier-plugin-tailwindcss": "^0.2.1",
     "tailwindcss": "^3.2.4",
     "ts-node": "^10.9.1",
-    "typescript": "^4.9.4",
-    "husky": "^8.0.0"
+    "typescript": "^4.9.4"
   },
   "private": true
 }
