+++ src/assets/adaptive-icon.png
